+++ db/migrate/20210820111745_add_important_to_notes.rb
@@ -1,0 +1,5 @@
+class AddImportantToNotes < ActiveRecord::Migration[6.0]
+  def change
+    add_column :notes, :important, :boolean
+  end
+end